+++ api/reviews.js
@@ -1,6 +1,7 @@
 "use strict";
 require('dotenv').config();
 const mysql = require('mysql2/promise');
+const jwt = require('jsonwebtoken');
 
 const mysql_host = process.env.MYSQL_HOST || 'localhost';
 const mysql_port = process.env.MYSQL_PORT || '3306';
@@ -22,11 +23,26 @@ const { validateAgainstSchema, extractValidFields } = require('../lib/validation
 
 exports.router = router;
 
+//adapted from challenge 6-1
+function requireAuthentication(req, res, next){
+    try {
+        const auth_value = req.get('Authorization');
+        const token = auth_value.split("[")[1].split("]")[0];
+        const payload = jwt.verify(token, process.env.JWT_SECRET_KEY);
+        // if we get here, success
+        req.userid=payload.userid;
+        req.admin=payload.admin;
+        next();
+    } catch(err) {
+        res.status(401).send(401);
+        // this means we failed
+    }
+};
+
 /*
  * Schema describing required/optional fields of a review object.
  */
 const reviewSchema = {
-  userid: { required: true },
   businessid: { required: true },
   dollars: { required: true },
   stars: { required: true },
@@ -37,9 +53,8 @@ const reviewSchema = {
 /*
  * Route to create a new review.
  */
-router.post('/', async function (req, res, next) {
+router.post('/', requireAuthentication, async function (req, res, next) {
   if (validateAgainstSchema(req.body, reviewSchema)) {
-
     const review = extractValidFields(req.body, reviewSchema);
 
     /*
@@ -53,7 +68,7 @@ router.post('/', async function (req, res, next) {
       });
     } else {
       let query = `insert into reviews (userid, businessid, dollars, stars, review) values\n`;
-      query += `("${review.userid}", "${review.businessid}", "${review.dollars}", "${review.stars}", "${review.review ?? ""}");`;
+      query += `("${req.userid}", "${review.businessid}", "${review.dollars}", "${review.stars}", "${review.review ?? ""}");`;
       await mysqlPool.query(query);
       res.status(201).json({
         id: review.id,
@@ -91,6 +106,9 @@ router.get('/:reviewID', async function (req, res, next) {
  */
 router.put('/:reviewID', async function (req, res, next) {
   const reviewID = parseInt(req.params.reviewID);
+  if (reviewID!=req.userid && !req.admin){
+    return res.status(403).send("unauthorized userid");
+  }
   const review_raw = await mysqlPool.query(`select * FROM reviews where id = ${reviewID}`);
   if (review_raw[0].length == 1) {
     if (validateAgainstSchema(req.body, reviewSchema)) {
@@ -130,6 +148,9 @@ router.put('/:reviewID', async function (req, res, next) {
  */
 router.delete('/:reviewID', async function (req, res, next) {
   const reviewID = parseInt(req.params.reviewID);
+  if (reviewID!=req.userid && !req.admin){
+    return res.status(403).send("unauthorized userid");
+  }
   const review_raw = await mysqlPool.query(`select * FROM reviews where id = ${reviewID}`);
   if (review_raw[0].length==1) {
     const success = await mysqlPool.query(`delete from reviews where id = ${reviewID}`);
